+++ gc/mark_and_compact/src/mark_and_compact.c
@@ -3,6 +3,7 @@
 #include <string.h>
 
 #include <mark_and_compact.h>
+#include <gc.h>
 #include <morecore.h>
 
 static void *gc_raw_alloc(size_t size);
