+++ gc/mark_and_compact/src/mark_and_compact.h
@@ -40,22 +40,6 @@ typedef struct {
 	int free_size;          // size from calculation
 } Heap_Info;
 
-/* Initialize a heap with a certain size for use with the garbage collector */
-extern void gc_init(int size);
-
-/* Announce you are done with the heap managed by the garbage collector */
-extern void gc_shutdown();
-
-/* Perform a mark_and_compact garbage collection, moving all live objects
- * to the start of the heap.
- */
-extern void gc();
-extern heap_object *gc_alloc(object_metadata *metadata, size_t size);
-extern void gc_add_root(void **p);
-extern void gc_mark();
-extern void gc_unmark();
-
-extern Heap_Info get_heap_info();
 
 #define gc_begin_func()		int __save = gc_num_roots()
 #define gc_end_func()		gc_set_num_roots(__save)
@@ -71,26 +55,6 @@ extern void gc_set_num_roots(int roots);
 extern void foreach_live(void (*action)(heap_object *));
 extern void foreach_object(void (*action)(heap_object *));
 extern bool ptr_is_in_heap(heap_object *p);
-
-static const size_t WORD_SIZE_IN_BYTES = sizeof(void *);
-static const size_t ALIGN_MASK = WORD_SIZE_IN_BYTES - 1;
-
-/* Pad size n to include header */
-static inline size_t size_with_header(size_t n) {
-	return n + sizeof(heap_object);
-}
-
-/* Align size n to nearest word size boundary (4 or 8) */
-static inline size_t align_to_word_boundary(size_t n) {
-	return (n & ALIGN_MASK) == 0 ? n : (n + WORD_SIZE_IN_BYTES) & ~ALIGN_MASK;
-}
-
-/* Convert a user request for n bytes into a size in bytes that has all necessary
- * header room and is word aligned.
- */
-static inline size_t request2size(size_t n) {
-	return align_to_word_boundary(size_with_header(n));
-}
 
 
 #ifdef __cplusplus
