+++ wlib/src/persistent_vector.c
@@ -26,9 +26,14 @@ SOFTWARE.
 #include <stdio.h>
 #include <math.h>
 
-#ifdef REFCOUNTING
-#elif MARK_AND_SWEEP
-#elif MARK_AND_COMPACT
+#if defined(MARK_AND_SWEEP)
+#include <mark_and_sweep.h>
+#include <gc.h>
+#elif defined(MARK_AND_COMPACT)
+#include <mark_and_compact.h>
+#include <gc.h>
+#elif defined(REFCOUNTING)
+#include <refcounting.h>
 #else // PLAIN
 typedef struct {} heap_object; // no extra header info needed
 #endif
